# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/parties:
    get:
      tags:
        - Parties
      summary: List Parties
      description: Get a paginated list of Parties with optional filtering and sorting
      operationId: listParties
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Parties
      summary: Create Party
      description: Create a new Party
      operationId: createParty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartyRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Party'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/party_phones:
    get:
      tags:
        - PartyPhones
      summary: List PartyPhones
      description: Get a paginated list of PartyPhones with optional filtering and sorting
      operationId: listPartyPhones
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyPhoneList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PartyPhones
      summary: Create PartyPhone
      description: Create a new PartyPhone
      operationId: createPartyPhone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartyPhoneRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyPhone'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Party:
      type: object
      required:
        - id
        - company_id
        - party_code
        - party_kind
        - name
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        party_code:
          type: string
          maxLength: 30
        party_kind:
          type: string
        name:
          type: string
          maxLength: 200
        legal_name:
          nullable: true
          type: string
          maxLength: 200
        first_name:
          nullable: true
          type: string
          maxLength: 100
        last_name:
          nullable: true
          type: string
          maxLength: 100
        npwp:
          nullable: true
          type: string
          maxLength: 25
        nik:
          nullable: true
          type: string
          maxLength: 20
        vat:
          nullable: true
          type: string
          maxLength: 32
        status:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    CreatePartyRequest:
      type: object
      required:
        - company_id
        - party_code
        - party_kind
        - name
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        party_code:
          type: string
          maxLength: 30
        party_kind:
          type: string
        name:
          type: string
          maxLength: 200
        legal_name:
          nullable: true
          type: string
          maxLength: 200
        first_name:
          nullable: true
          type: string
          maxLength: 100
        last_name:
          nullable: true
          type: string
          maxLength: 100
        npwp:
          nullable: true
          type: string
          maxLength: 25
        nik:
          nullable: true
          type: string
          maxLength: 20
        vat:
          nullable: true
          type: string
          maxLength: 32
        status:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    PartyList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Party'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PartyPhone:
      type: object
      required:
        - id
        - party_id
        - company_id
        - label
        - phone
        - is_primary
        - metadata
      properties:
        id:
          type: string
          format: uuid
        party_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        label:
          type: string
          maxLength: 30
        phone:
          type: string
          maxLength: 30
        is_primary:
          type: boolean
        metadata:
          type: object
...
    CreatePartyPhoneRequest:
      type: object
      required:
        - party_id
        - company_id
        - label
        - phone
        - is_primary
        - metadata
      properties:
        party_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        label:
          type: string
          maxLength: 30
        phone:
          type: string
          maxLength: 30
        is_primary:
          type: boolean
        metadata:
          type: object
...
    PartyPhoneList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PartyPhone'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []